crush: CrushTester: add test_with_crushtool()
Instead of testing directly using CrushTester::test(), which may break
the monitor if a malformed map is passed and we don't manage to catch
it, fork and run crushtool instead.  This way, if crushtool segfaults,
the monitor (or whomever the call might be) will be protected from a
catastrophic failure.

Signed-off-by: Joao Eduardo Luis <joao@redhat.com>